backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.30.5"
python-decouple = "^3.8"
instructor = "^1.3.2"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
qdrant-client = "^1.9.1"
hubspot-api-client = "^9.0.0"
requests = "^2.32.3"
pandas = "^2.2.2"
anthropic = "^0.31.2"
sse-starlette = "^2.1.2"
psycopg2 = "^2.9.9"
google-generativeai = "^0.8.4"
groq = "^0.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
